refactor: 패키지 구조 변경 (#30)
기존 favorite 디렉토리 안에 있던 서비스 및 implement 패키지를 user 디렉토리 안의 implement/mapping 디렉토리와 userService 안으로 이동시켰다.
