+++ src/main/java/com/danthis/backend/application/user/implement/mapping/UserDancerReader.java
@@ -1,9 +1,12 @@
 package com.danthis.backend.application.user.implement.mapping;
 
+import com.danthis.backend.domain.mapping.userdancer.UserDancer;
 import com.danthis.backend.domain.mapping.userdancer.repository.UserDancerRepository;
 import com.danthis.backend.domain.user.User;
 import java.util.List;
 import lombok.RequiredArgsConstructor;
+import org.springframework.data.domain.Page;
+import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Component;
 
 @Component
@@ -17,5 +20,9 @@ public class UserDancerReader {
                                .stream()
                                .map(userDancer -> userDancer.getDancer().getId())
                                .toList();
+  }
+
+  public Page<UserDancer> readDancersByUserId(Long userId, Pageable pageable) {
+    return userDancerRepository.findByUserId(userId, pageable);
   }
 }
