+++ src/main/java/com/danthis/backend/application/user/implement/mapping/WishListReader.java
@@ -1,7 +1,5 @@
-package com.danthis.backend.application.favorite.implement;
+package com.danthis.backend.application.user.implement.mapping;
 
-import com.danthis.backend.domain.mapping.userdancer.UserDancer;
-import com.danthis.backend.domain.mapping.userdancer.repository.UserDancerRepository;
 import com.danthis.backend.domain.mapping.wishlist.WishList;
 import com.danthis.backend.domain.mapping.wishlist.repository.WishListRepository;
 import lombok.RequiredArgsConstructor;
@@ -11,14 +9,9 @@ import org.springframework.stereotype.Component;
 
 @Component
 @RequiredArgsConstructor
-public class FavoriteReader {
+public class WishListReader {
 
-  private final UserDancerRepository userDancerRepository;
   private final WishListRepository wishListRepository;
-
-  public Page<UserDancer> readDancersByUserId(Long userId, Pageable pageable) {
-    return userDancerRepository.findByUserId(userId, pageable);
-  }
 
   public Page<WishList> readWishListByUserId(Long userId, Pageable pageable) {
     return wishListRepository.findByUserId(userId, pageable);
